+++ src/dialog/_example/base.vue
@@ -4,8 +4,7 @@
     <t-dialog
       v-model:visible="visible"
       header="对话框标题"
-      body="对话框内容"
-      attach="body"
+      width="40%"
       :confirm-on-enter="true"
       :on-cancel="onCancel"
       :on-esc-keydown="onEscKeydown"
@@ -13,13 +12,23 @@
       :on-overlay-click="onOverlayClick"
       :on-close="close"
       :on-confirm="onConfirmAnother"
-    />
+    >
+      <t-space direction="vertical" style="width: 100%">
+        <div>
+          <p>这是弹框内容</p>
+          <p>This is Dialog Content</p>
+        </div>
+        <t-pagination v-model="current" v-model:pageSize="pageSize" :total="30" />
+      </t-space>
+    </t-dialog>
   </t-space>
 </template>
 <script setup>
 import { ref } from 'vue';
 
 const visible = ref(false);
+const current = ref(1);
+const pageSize = ref(10);
 const onClick = (context) => {
   console.log('点击了确认按钮，弹出弹窗', context);
   visible.value = true;