+++ src/dialog/plugin.tsx
@@ -89,7 +89,11 @@ const createDialog: DialogMethod = (props: DialogOptions) => {
       visible.value = false;
       setTimeout(() => {
         dialogComponent.unmount();
+        wrapper.remove();
       }, 300);
+    },
+    setConfirmLoading: (val: boolean) => {
+      dialog.update({ confirmLoading: val });
     },
   };
   return dialogNode;
@@ -112,9 +116,9 @@ const extraApi: ExtraApi = {
   alert,
 };
 
-export type DialogPluginType = Plugin & ExtraApi & DialogAlertMethod;
+export type DialogPluginType = Plugin & ExtraApi & DialogMethod;
 
-export const DialogPlugin: DialogPluginType = createDialog as DialogPluginType;
+export const DialogPlugin = createDialog as DialogPluginType;
 
 DialogPlugin.install = (app: App): void => {
   app.config.globalProperties.$dialog = createDialog;