+++ src/table/hooks/useFilter.tsx
@@ -1,11 +1,14 @@
-import { toRefs, ref, watch, computed, SetupContext } from 'vue';
+import { toRefs, ref, watch, computed, SetupContext, h } from 'vue';
 import useClassName from './useClassName';
 import TButton from '../../button';
-import { TdPrimaryTableProps, PrimaryTableCol, TableRowData, FilterValue } from '../type';
+import { TdPrimaryTableProps, PrimaryTableCol, TableRowData, FilterValue, TableFilterChangeContext } from '../type';
 import useDefaultValue from '../../hooks/useDefaultValue';
 import { useTNodeDefault } from '../../hooks/tnode';
 import TableFilterController from '../filter-controller';
 import { useConfig } from '../../hooks/useConfig';
+import isFunction from 'lodash/isFunction';
+import { getColumnsResetValue } from '../../_common/js/table/utils';
+import { renderTitle } from './useTableHeader';
 
 function isFilterValueExist(value: any) {
   const isArrayTrue = value instanceof Array && value.length;
@@ -28,9 +31,9 @@ function filterEmptyData(data: FilterValue) {
 
 export default function useFilter(props: TdPrimaryTableProps, context: SetupContext) {
   const primaryTableRef = ref(null);
-  const { t, globalConfig } = useConfig('table');
+  const { t, globalConfig } = useConfig('table', props.locale);
   const renderTNode = useTNodeDefault();
-  const { filterValue } = toRefs(props);
+  const { filterValue, columns } = toRefs(props);
   const { tableFilterClasses, isFocusClass } = useClassName();
   const isTableOverflowHidden = ref<boolean>();
 
@@ -83,7 +86,7 @@ export default function useFilter(props: TdPrimaryTableProps, context: SetupCont
     getAllColumns(props.columns, columns);
     columns
       .filter((col) => col.filter)
-      .forEach((col) => {
+      .forEach((col, index) => {
         let value = tFilterValue.value[col.colKey];
         if (col.filter.list && !['null'].includes(String(value))) {
           const formattedValue = value instanceof Array ? value : [value];
@@ -96,7 +99,9 @@ export default function useFilter(props: TdPrimaryTableProps, context: SetupCont
           value = label.join();
         }
         if (isFilterValueExist(value)) {
-          arr.push(`${col.title}：${value}`);
+          const label = isFunction(col.filter?.label) ? col.filter.label(h) : col.filter?.label;
+          const title = renderTitle(context.slots, col, index);
+          arr.push(`${label || title}：${value}`);
         }
       });
     return arr.join('；');
@@ -118,12 +123,16 @@ export default function useFilter(props: TdPrimaryTableProps, context: SetupCont
     };
     innerFilterValue.value = filterValue;
     if (!column.filter.showConfirmAndReset) {
-      emitFilterChange(filterValue, column);
+      emitFilterChange(filterValue, 'filter-change', column);
     }
   }
 
-  function emitFilterChange(filterValue: FilterValue, column?: PrimaryTableCol) {
-    setTFilterValue(filterValue, { col: column });
+  function emitFilterChange(
+    filterValue: FilterValue,
+    trigger: TableFilterChangeContext<TableRowData>['trigger'],
+    column?: PrimaryTableCol,
+  ) {
+    setTFilterValue(filterValue, { col: column, trigger });
     props.onChange?.({ filter: filterValue }, { trigger: 'filter' });
   }
 
@@ -139,15 +148,16 @@ export default function useFilter(props: TdPrimaryTableProps, context: SetupCont
         }[column.filter.type] ??
         '',
     };
-    emitFilterChange(filterValue, column);
+    emitFilterChange(filterValue, 'reset', column);
   }
 
   function onResetAll() {
-    emitFilterChange({}, undefined);
+    const resetValue = getColumnsResetValue(columns.value);
+    emitFilterChange(resetValue, 'clear', undefined);
   }
 
   function onConfirm(column: PrimaryTableCol) {
-    emitFilterChange(innerFilterValue.value, column);
+    emitFilterChange(innerFilterValue.value, 'confirm', column);
   }
 
   // 图标：内置图标，组件自定义图标，全局配置图标
@@ -164,6 +174,7 @@ export default function useFilter(props: TdPrimaryTableProps, context: SetupCont
         isFocusClass={isFocusClass}
         popupProps={col.filter.popupProps}
         attach={props.attach}
+        locale={props.locale}
         onReset={onReset}
         onConfirm={onConfirm}
         onInnerFilterChange={onInnerFilterChange}